+++ CHANGELOG.md
@@ -3,11 +3,16 @@
 Note: the BigBuy API is not versioned and some endpoints are often added or removed without notice.
 The removal of methods that call deleted endpoints is not considered a breaking change.
 
+## 3.21.1 (unreleased)
+
+* Add `TypedDict`s for the return hint of select endpoints: `BBProductCategoryDict`, `BBProductInformationDict`
+* Fix the return type hint of `get_product_categories`
+
 ## 3.21.0 (2024/11/29)
 
 * Drop support for Python 3.8
-* Add typed dicts for select endpoints: `BBProductImagesDict`, `BBTaxonomyDict`, `BBProductTaxonomyDict`,
-  `BBLowestShippingCostDict`, `BBAttributeDict`, `BBAttributeGroupDict`
+* Add `TypedDict`s for the return hint of select endpoints: `BBProductImagesDict`, `BBTaxonomyDict`,
+  `BBProductTaxonomyDict`, `BBLowestShippingCostDict`, `BBAttributeDict`, `BBAttributeGroupDict`
 * Add explicit support for Python 3.13
 
 ## 3.20.1 (2024/08/01)